requirements.txt
groq
webwhatsapp-wrapper
python-dotenv
Pillow
requests
selenium